refactor: mv fetchErc20Decimals to utils/token
