+++ ui/pages/confirmations/constants/token.ts
@@ -1,0 +1,1 @@
+export const ERC20_DEFAULT_DECIMALS = 18;
